requirements.txt
certifi==2022.6.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
pipreqs==0.4.11
PyYAML==6.0
requests==2.28.1
tgt==1.4.4
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.8.1
